Correct murano-agent reference to message headers
The python-murano agent refers to a 'header'
attribute when trying to extract the message
id from the list of headers.

The message object is of the class
kombu.transport.pyamqp.Message, which does not
have a 'header' attribute, but a 'headers'
attribute.

The reference needs to be corrected to 'headers'

Change-Id: Ibf470090e01d159cdfea13d142f6b712c649183a
Closes-Bug: #1315571