Remove hardcoded version
Generally everything other than API servers uses semantic versioning,
not pbr's pre-version facility. This line causes local package builds
to have the wrong version today (0.3.dev when 2014.2 is the version
under development), and pbr will in future detect this and error. If
Murano wants to use preversioning for the agent, this can instead be
changed to

  version = 2014.2

Change-Id: If65a92f8673faefd77419354f24ba38c0c18c05f