Fix Mutable defalut argument
Python’s default arguments are evaluated once when the
function is defined, not each time the function is called.
This means that if you use a mutable default argument
and mutate it, you will and have mutated that object for
all future calls to the function as well.

more details about this wrong usage here:
http://docs.python-guide.org/en/latest/writing/gotchas/#mutable-default-arguments

Change-Id: Ic8eb1328852eb8a3180a08466bf18b72fd1f7956