RabbitMQ heartbeats for consumer threads
Sometimes network connection between agent and RabbitMQ
becomes corrupted without agent being aware. Usually
it happen due to some networking reconfiguration. For example
changes to iptables. Because consumer connection only reads from
the socket it cannot detect broken connection and will wait
for the message forever.

This commit enables periodic heartbeats that agent sends to the
server. If RabbitMQ doesn't receive heartbeats for some time it
disconnects the client. On the other hand because heartbeat
is a write operation it lets the client to detect broken connection
as well.

Change-Id: Id326f8d3b6dce04d83472837e11342efa8eb32fe
Closes-Bug: #1583007
Partially-Implements: blueprint rabbitmq-heartbeats