Adds TLS/SSL Version Support to Murano Agent.
Already implemented for Murano, but need Murano Agent
support as well.

Currently, Murano supports enabling SSL in Murano Engine [0],
but does not have a param for the user to specify the
desired SSL version. This is important because some
versions of SSL are less secure than others [1].

This patch adds ssl_version support to Murano Agent.

[0] https://docs.openstack.org/developer/murano/administrator-guide/deploy_murano/configure_ssl.html
[1] https://www.wolfssl.com/wolfSSL/Blog/Entries/2010/10/7_Differences_between_SSL_and_TLS_Protocol_Versions.html

Change-Id: I410a38d3c8294ce63c15ae98f627fa5c69a3d118
Partially-Implements: blueprint add-tls-support
Depends-On: I71c36c455cde658f402a19c59d7966cee8544cf1